Add GUI script that counts tokens in a given text.
Some people are not going to want to run a CLI program to count tokens,
but prefer a GUI application where they can paste and count tokens.

The goal of the script is to be as portable and robust as possible.

Use `tkinter`, which is free and a part of the standard library.

Very little effort has gone into making this pretty (read no effort).
It's basically just a proof of concept which I can share with my
non-technical friends/colleagues.